Refactor http transport options
